+++ Changelog.md
@@ -28,6 +28,7 @@ Bugfixes:
  * SMTChecker: Fix internal error when applying arithmetic operators to fixed point variables.
  * SMTChecker: Fix internal error when short circuiting Boolean expressions with function calls in state variable initialization.
  * SMTChecker: Fix internal error when assigning to index access inside branches.
+ * SMTChecker: Fix internal error on try/catch clauses with parameters.
 
 ### 0.6.8 (2020-05-14)